Move Polygon3D to shared project
