+++ Source/HelixToolkit.Shared/HelixToolkit.Shared.projitems
@@ -13,6 +13,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Geometry\MeshBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Geometry\MeshGeometryHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Geometry\Polygon.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Geometry\Polygon3D.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Geometry\SharedFunctions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Geometry\SweepLinePolygonTriangulator.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HelixToolkitException.cs" />
